API: Delay string and number promotion deprecation/future warning
Unfortunately, this seems to cause some pain in pandas at least
to circumvent the warning.

Delaying may complicate the creation of ufuncs for strings, but
otherwise should not be a big problem.  We have to make sure that
we can reimplement the change quickly though, it is an important
change in my opinion.