Unpin pydata sphinx theme and update config.
As of v0.6, the pydata-sphinx-theme has a collapsible navigation
feature for the sidebar that significantly increases the build time
and html file size for the docs.
We had originally pinned to 0.5.2 to avoid this, but the feature is
now configurable with the collapse_navigation=True value.
Remove the upper-bound pinning the theme and instead use the
configuration option to keep doc build times reasonable.