BUG: ensure np.median does not drop subclass for NaN result.
Currently, np.median is almost completely safe for subclasses, except
if the result is NaN.  In that case, it assumes the result is a scalar
and substitutes a NaN with the right dtype.  This PR fixes that, since
subclasses like astropy's Quantity generally use array scalars to
preserve subclass information such as the unit.