MAINT: Fix LGTM.com error
Unmatchable caret in regular expression
This regular expression includes an unmatchable caret at offset 18.

Indeed this regex couldn't possibly match anything. Not only does the
regex look obsolete, attempting to match "inline" or "static" specifiers
that I cannot find in the sources produced by 'f2c', but the very logic
of function removeSubroutinePrototypes() makes it impossible to handle
'/* Subroutine */' declarations that span multiple lines.

All in all, function removeSubroutinePrototypes() does nothing at all.
We choose not to change that, because it works and we have no real hint
what it should do and why. We just document that.