DOC: Mention `nan` results in `power` and `float_power`.
By design, `power` and `float_power` return `nan` when a
negative value is raised to a non-integral value.  There have
been several issues raised about this over the last few years,
so let's explicitly mention this in the docstrings, and give an
example of explicitly specifying `dtype=complex` to get complex
results intead of `nan`.

Closes gh-20186.
Closes gh-12182.
Closes gh-12118.