MAINT,TST: Avoid small positive integers in refcount test
This hopefully makes the test less susceptible to errors, but
we have to use small "singleton" integers for the test to work.

It is not clear that this will actually fix the issue always.
If the issue crops up again, we should simply remove all manual
asserts and rely on `pytest-leaks` as a leak checker.

Closes gh-17381