MAINT: update OpenBLAS to 0.3.18
Not sure what the usual timeline on updating this is.
Running the CI on this PR is one way to narrow down the Cygwin problems addressed in PR#20654.
If this build also has four failures, then OpenBLAS is probably the problem.
If this build uses the new OpenBLAS with no problems, then I need to see what else changed
in the Cygwin packages around the time the builds started failing.