BUG: Fix CI failure by adding symbols to be exported on Cygwin
There are two types of Cython in the Cygwin environment.
One is a Cygwin package's Cython,
which does not add a `__declspec(dllexport)` attribute to symbols
due to its custom patch.
The other is Cython, which is installed with pip and has no custom patches,
so the symbols are added the `__declspec(dllexport)` attribute.

Since CI uses symbols in distributions.c by cffi, we need to export them.
The commit b797a0cb97a8a9e7e36962dc8c37c075967aa88b exports the symbols
by adding the `__declspec(dllexport)` attribute to them.
However, in the Cygwin package's Cython environment,
the symbols required for import are not exported
and the module cannot be imported.
https://cygwin.com/pipermail/cygwin/2022-January/250481.html

By the previous commit, the `__declspec(dllexport)` attribute is removed,
so in the environment, the required symbols are exported
and the module can be imported.
However, on the other hand, in the pip-installed Cython environment,
the symbols used by cffi are not exported, so CI failed.

This commit exports the symbols without `__declspec(dllexport)`
for use by cffi.
It exports both the symbols required for import
and the symbols used for cffi,
regardless of which Cython environment is used.