DOC: disambiguate :: in rst. (#21037)
Having the `::` on new line can be ambiguous for RST parsers,
as `:` is a valid character for header underlines. And as underlines do
not have to be as long as the title for some rst parser this appears to
be a title.

Workaround is to have either a blank line, or put the `::` at the end of
previous one.