BUG,MAINT: Never use negative offset and tighten promote-types
Negative offsets in structured dtypes give indirect access outside
of the original data in principle, thus we should always reject
them (this is as of now practically unused in NumPy).

Further, `np.result_type` and `np.promote_types` should behave
the same as canonicalization for twice the same input (at least
mostly, metadata is a rare exception).

This was not the case, because promote-types had a fast path that
would return the wrong thing for structured dtypes.
Thus, reject them from the fast-path, and limit it to builtin types
for now.