NEP: Add early draft for fixing scalar promotion rules
Note, this really is a very early draft.  For myself, I have not
even 100% made up my mind that this is actually the best approach.

However, to solicitate feedback and input, I would like to put it
out there (and have a nice rendered version).

For those mainly interested in the possible different approaches
(i.e. making decisions), the "Alternatives" section is probably most
interesting.

Please make a note of blatend errors or unnecessary repitition, but
don't be too surprised if you see them, I will not be surprised if
this needs to be redone from scratch eventually.