DOC: Note interop from "subclassing" docs and explain when to avoid
I think we have consensus that for many users subclassing is a bad idea,
but our docs don't really give much of a hint as to why, and what else.

We long have the answers for these, so this is a start to actually
write them down.

Addresses some of the points in gh-20998.